realized logout

+++ services/contacts.js
@@ -1,24 +1,30 @@
 const Contact = require('../model/schemas/contact');
 
-async function listContacts() {
-  const results = await Contact.find({});
+async function listContacts(userId) {
+  const results = await Contact.find({ owner: userId }).populate({
+    path: 'owner',
+    select: 'name email sex -_id',
+  });
   return results;
 }
-async function getById(id) {
-  const results = await Contact.findOne({ _id: id });
+async function getById(id, userId) {
+  const results = await Contact.findOne({ _id: id, owner: userId }).populate({
+    path: 'owner',
+    select: 'name email sex -_id',
+  });
   return results;
 }
-async function removeContact(id) {
-  const result = await Contact.findByIdAndRemove({ _id: id });
+async function removeContact(id, userId) {
+  const result = await Contact.findByIdAndRemove({ _id: id, owner: userId });
   return result;
 }
 async function addContact(body) {
   const result = await Contact.create(body);
   return result;
 }
-async function updateContact(id, body) {
+async function updateContact(id, body, userId) {
   const result = await Contact.findByIdAndUpdate(
-    { _id: id },
+    { _id: id, owner: userId },
     { ...body },
     { new: true },
   );
